requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.2
google-generativeai==0.7.2
apscheduler==3.10.4
razorpay==1.4.2
